ModelSerializer refactoring part 2

+++ utils/ModelSerializer/src/ObjProperties.h
@@ -10,10 +10,10 @@ struct ObjProperties
 	{
 	}
 
-	UINT32 vertexCount;
-	UINT32 normalCount;
-	UINT32 uvCount;
-	UINT32 faceCount;
+	UINT64 vertexCount;
+	UINT64 normalCount;
+	UINT64 uvCount;
+	UINT64 faceCount;
 	float maxVertexPos;
 };
